+++ src/azure-cli/azure/cli/command_modules/cosmosdb/custom.py
@@ -5,6 +5,7 @@
 
 # pylint: disable=too-many-lines
 
+from enum import Enum
 from knack.log import get_logger
 from knack.util import CLIError
 
@@ -28,6 +29,13 @@ from azure.mgmt.cosmosdb.models import (
 )
 
 logger = get_logger(__name__)
+
+
+class CosmosKeyTypes(Enum):
+    keys = "keys"
+    read_only_keys = "read-only-keys"
+    connection_strings = "connection-strings"
+
 
 DEFAULT_INDEXING_POLICY = """{
   "indexingMode": "consistent",
@@ -233,6 +241,17 @@ def cli_cosmosdb_list(client, resource_group_name=None):
         return client.list_by_resource_group(resource_group_name)
 
     return client.list()
+
+
+# pylint: disable=line-too-long
+def cli_cosmosdb_keys(client, resource_group_name, account_name, key_type=CosmosKeyTypes.keys.value):
+    if key_type == CosmosKeyTypes.keys.value:
+        return client.list_keys(resource_group_name, account_name)
+    if key_type == CosmosKeyTypes.read_only_keys.value:
+        return client.list_read_only_keys(resource_group_name, account_name)
+    if key_type == CosmosKeyTypes.connection_strings.value:
+        return client.list_connection_strings(resource_group_name, account_name)
+    raise CLIError("az cosmosdb keys list: '{0}' is not a valid value for '--type'. See 'az cosmosdb keys list --help'.".format(key_type))
 
 
 def cli_cosmosdb_sql_database_create(client,